issue-53: code-review third part of the comments
